requirements.txt

django
django-tastypie
uuidfield
-e git://github.com/Adys/django-push-notifications.git#egg=push_notifications